fixup! Move models into subclasses
