+++ pumaguard/classify.py
@@ -8,7 +8,7 @@ import argparse
 import logging
 
 from pumaguard.model import (
-    Model,
+    model_factory,
 )
 from pumaguard.presets import (
     Preset,
@@ -39,7 +39,7 @@ def main(options: argparse.Namespace, presets: Preset):
     """
 
     logger.debug('loading model from %s', presets.model_file)
-    model = Model(presets).model
+    model = model_factory(presets).model
 
     for image in options.image:
         prediction = classify_image(presets, model, image)
