+++ pumaguard/server.py
@@ -10,7 +10,7 @@ import subprocess
 import threading
 import time
 
-from pumaguard.model import (
+from pumaguard.model_factory import (
     model_factory,
 )
 from pumaguard.presets import (
